Tests: Added tests for the DataResource Class
Since this class does most of the business logic for US1 and US2 it is the only thing being tested this iteration.
What is missing? Test for saveDb as it is not written yet.

These tests will change and be more effective during the next iteration after the major refactor happens.